Updating setup script
Add check to see if .tmux.conf already exists and don't overwrite it.

+++ setup.sh
@@ -3,6 +3,7 @@
 dir=$PWD
 
 # pull in vim color
+echo "pull in gruvbox.vim color"
 cd ~
 
 if [ ! -d ".vim/colors" ]; then
@@ -12,11 +13,14 @@ fi
 cd .vim/colors
 curl -O https://raw.githubusercontent.com/morhetz/gruvbox/master/colors/gruvbox.vim
 
-# pull down tmux setup
-cd ~
-git clone https://github.com/gpakosz/.tmux.git
-ln -s -f .tmux/.tmux.conf
-cp .tmux/.tmux.conf.local .
+# pull down tmux setup if it doesn't exist
+if [ ! -f "~/.tmux.conf" ]; then
+  echo ".tmux.conf does not exist, pulling down config"
+  cd ~
+  git clone https://github.com/gpakosz/.tmux.git
+  ln -s -f .tmux/.tmux.conf
+  cp .tmux/.tmux.conf.local .
+fi
 
 # copy over dot files
 cd $dir
